Moved Events folder to root better info on exception allow overwrite zip
