+++ PLang/Container/Container.cs
@@ -8,8 +8,8 @@ using Nethereum.Web3;
 using Nethereum.Web3.Accounts;
 using Nostr.Client.Client;
 using PLang.Building;
-using PLang.Building.Events;
 using PLang.Building.Parsers;
+using PLang.Events;
 using PLang.Exceptions;
 using PLang.Exceptions.AskUser;
 using PLang.Exceptions.Handlers;
@@ -41,7 +41,7 @@ using static PLang.Modules.DbModule.ModuleSettings;
 
 namespace PLang.Container
 {
-	public static class Instance
+    public static class Instance
 	{
 		private static readonly string PrefixForTempInjection = "__Temp__";
 		public record InjectedType(string InjectorName, Type ServiceType, Type ImplementationType);
@@ -469,7 +469,9 @@ namespace PLang.Container
 
 			if (!fileSystem.Directory.Exists(".services"))
 			{
-				throw new RuntimeException($".services folder not found in {fileSystem.RootDirectory}");
+				var logger = container.GetInstance<ILogger>();
+				logger.LogWarning($".services folder not found in {fileSystem.RootDirectory}. If you have modified injection you need to Delete the file Start/00. Goal.pr or events/00. Goal.pr. Will be fixed in future release.");
+				return null;
 			}
 
 			string dllFilePath = Path.GetDirectoryName(Path.Combine(fileSystem.GoalsPath, ".services", injectorType));
@@ -480,7 +482,9 @@ namespace PLang.Container
 				var moduleFolderPath = Path.Combine(fileSystem.GoalsPath, ".services", dllFilePath);
 				if (!fileSystem.Directory.Exists(moduleFolderPath))
 				{
-					throw new RuntimeException($"{injectorType} injection folder could not be found. Path {moduleFolderPath}");
+					var logger = container.GetInstance<ILogger>();
+					logger.LogWarning($"{injectorType} injection folder could not be found. Path {moduleFolderPath}");
+					return null;
 				}
 
 
@@ -524,7 +528,8 @@ namespace PLang.Container
 			if (implementationType == null)
 			{
 				logger.LogError($"ERROR: implementationType is null for {injectorType} - pathToModule:{pathToModule}");
-				throw new RuntimeException($"Loading '{injectorType}': interface:{injectorType} | type is:{implementationType} | reservedKeyword:{reservedKeyword} | isGlobalForApp:{isGlobalForApp} | pathToModule:{pathToModule}");
+				return;
+				//throw new RuntimeException($"Loading '{injectorType}': interface:{injectorType} | type is:{implementationType} | reservedKeyword:{reservedKeyword} | isGlobalForApp:{isGlobalForApp} | pathToModule:{pathToModule}");
 			}
 			if (container.CanGetInstance(interfaceType, implementationType.FullName)) return;
