+++ PLang/Exceptions/GoalNotFoundException.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PLang.Building.Model;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -10,6 +11,7 @@ namespace PLang.Exceptions
 	{
 		public string AppPath { get; private set; }
 		public string GoalName { get; private set; }
+
 		public GoalNotFoundException(string message, string appPath, string goalName) : base(message) { 
 			this.AppPath = appPath;
 			this.GoalName = goalName;
