+++ PLang/Exceptions/RuntimeEventException.cs
@@ -1,5 +1,5 @@
-﻿using PLang.Building.Events;
-using PLang.Building.Model;
+﻿using PLang.Building.Model;
+using PLang.Events;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace PLang.Exceptions
 {
-	public class RuntimeEventException : Exception
+    public class RuntimeEventException : Exception
 	{
 		EventBinding? eventModel;
 		public RuntimeEventException(string message, EventBinding? eventModel = null) : base(message) {
