+++ PLang/Interfaces/IArchiver.cs
@@ -4,8 +4,8 @@ namespace PLang.Interfaces
 {
 	public interface IArchiver
 	{
-		Task CompressFiles(string[] filePaths, string saveToPath, int compressionLevel = 0);
-		Task CompressDirectory(string sourceDirectoryName, string destinationArchiveFileName, int compressionLevel = 0, bool includeBaseDirectory = true);
+		Task CompressFiles(string[] filePaths, string saveToPath, int compressionLevel = 0, bool overwrite = false);
+		Task CompressDirectory(string sourceDirectoryName, string destinationArchiveFileName, int compressionLevel = 0, bool includeBaseDirectory = true, bool overwrite = false);
 		Task DecompressFile(string sourceArchiveFileName, string destinationDirectoryName, bool overwrite = false);
 	}
 }
