+++ PLang/Modules/OutputModule/Builder.cs
@@ -1,0 +1,15 @@
+﻿using PLang.Building.Model;
+
+namespace PLang.Modules.OutputModule
+{
+	public class Builder : BaseBuilder
+	{
+		public override async Task<Instruction> Build(GoalStep step)
+		{
+
+			return await base.Build(step);
+			
+
+		}
+	}
+}
