+++ PlangTests/Building/Events/EventBuilderTests.cs
@@ -4,6 +4,7 @@ using Newtonsoft.Json;
 using NSubstitute;
 using PLang.Building.Model;
 using PLang.Building.Parsers;
+using PLang.Events;
 using PLang.Exceptions;
 using PLang.Interfaces;
 using PLang.Models;
